requirements-dev.txt
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
filelock==3.18.0
hypothesis==6.137.3
identify==2.6.13
iniconfig==2.1.0
mypy==1.17.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
ruff==0.12.8
# Editable Git install with no remote (sierra-sync==0.0.1)
-e c:\dev\datamatching
sortedcontainers==2.4.0
types-PyYAML==6.0.12.20250809
typing_extensions==4.14.1
virtualenv==20.33.1
